refactor: CartPage 리팩토링
- 가독성을 위해 빈 장바구니 페이지 다른 파일로 분리
- 사용 안하는 타입 및 주석 제거
